% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{complier_mod}
\alias{complier_mod}
\title{Train complier model using ensemble methods}
\usage{
complier_mod(data, ID = NULL, SL.library = NULL)
}
\arguments{
\item{data}{list object of experimental data.}

\item{ID}{string for name of indentifier variable.}

\item{SL.library.}{Employs extreme gradient boosting, elastic net regression,
random forest, and neural nets.}
}
\value{
model object of trained model.
}
\description{
Train model using group exposed to treatment with compliance as binary
outcome variable and covariates.
}
