% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{create.SL.gam}
\alias{create.SL.gam}
\title{Title}
\usage{
create.SL.gam(deg.gam = c(3, 4))
}
\arguments{
\item{deg.gam}{}
}
\description{
Creates gam wrappers in the global environment with different degrees.
The default value for deg.gam in SL.gam is 2
}
\keyword{internal}
