% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{create.SL}
\alias{create.SL}
\title{create.SL}
\usage{
create.SL(learners = "all")
}
\arguments{
\item{learners}{create ML wrappers for ensemble methods}
}
\description{
create.SL
}
\keyword{internal}
