% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{define.SL.reg.library}
\alias{define.SL.reg.library}
\title{define.SL.reg.library}
\usage{
define.SL.reg.library(
  SL.library.reg = c("SL.gam", "SL.gam.3", "SL.gam.4", "SL.gbm.gaussian", "SL.glmnet",
    "SL.glmnet.0", "SL.glmnet.0.25", "SL.glmnet.0.5", "SL.glmnet.0.75",
    "SL.randomForest.4", "SL.randomForest.6")
)
}
\arguments{
\item{SL.library.reg}{}
}
\description{
define.SL.reg.library
}
\keyword{internal}
