% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_neural.R
\name{neuralnet_predict}
\alias{neuralnet_predict}
\title{Predicting Compliance from experimental data}
\usage{
neuralnet_predict(neuralnet.complier.mod, exp.data, treat.var, compl.var)
}
\arguments{
\item{neuralnet.complier.mod}{results from \code{neuralnet_complier_mod}}

\item{exp.data}{\code{data.frame} of experimental data}

\item{treat.var}{string for treatment variable}

\item{compl.var}{string for compliance variable}
}
\value{
\code{data.frame} object with true compliers, predicted compliers in the
control group, and all compliers (actual + predicted).
}
\description{
Predicting Compliance from control group experimental data
}
