% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{popcall}
\alias{popcall}
\title{create list object of population data for easy data processing}
\usage{
popcall(
  response.formula,
  compl.var,
  data,
  weights = NULL,
  cluster = NULL,
  ID = NULL
)
}
\arguments{
\item{ID}{}
}
\description{
create list object of population data for easy data processing
}
\keyword{internal}
