% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearners.R
\name{predict.SL.mean}
\alias{predict.SL.mean}
\title{Title}
\usage{
\method{predict}{SL.mean}(object, newdata, family, X = NULL, Y = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\keyword{internal}
