% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patt_ensemble.R
\name{response_model}
\alias{response_model}
\title{Response model from experimental data using SL ensemble}
\usage{
response_model(
  exp.data,
  exp.compliers,
  family = "binomial",
  ID = NULL,
  SL.library = NULL
)
}
\arguments{
\item{exp.data}{experimental dataset.}

\item{exp.compliers}{\code{data.frame} object of compliers from
\code{complier_predict}.}

\item{family}{string for \code{"gaussian"} or \code{"binomial"}.}

\item{ID}{string for identifier variable.}

\item{SL.library}{vector of names of ML algorithms used for ensemble model.}
}
\value{
trained response model.
}
\description{
Train response model (response variable as outcome and covariates) from all
compliers (actual + predicted) in experimental data using SL ensemble.
}
