% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtc.R
\name{WtC}
\alias{WtC}
\title{WtC}
\usage{
WtC(
  x,
  y,
  c = NULL,
  weight = NULL,
  weighty = NULL,
  weightc = NULL,
  cluster = NULL,
  clustery = NULL,
  clusterc = NULL,
  samedata = TRUE,
  alternative = "two.tailed",
  mean1 = TRUE,
  bootse = TRUE,
  bootp = FALSE,
  bootn = 1000,
  drops = "pairwise",
  equivalence = FALSE
)
}
\arguments{
\item{equivalence}{}
}
\description{
Weighted t-test to estimate PATT-C
}
